requirements.txt
httpx>=0.27.0
pydantic>=2.7.0
websockets>=12.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.1
